[api] Use unsigned length values in V8::String Value types
This CL switches lengths of V8::String::Value and V8::String::ValueView
to uint32_t and the length of V8::String::Utf8Value to size_t. This is
both a follow-up for https://crrev.com/c/5332375 (which converted the
internal String::length_ to an uint32_t) and preparation for
https://crrev.com/c/5975682 which will introduce a new API for Utf8
encoding a String that uses size_t.

Bug: 369748453, 373485796
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6012365
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97111}

+++ include/v8-primitive.h
@@ -558,7 +558,7 @@ class V8_EXPORT String : public Name {
     ~Utf8Value();
     char* operator*() { return str_; }
     const char* operator*() const { return str_; }
-    int length() const { return length_; }
+    size_t length() const { return length_; }
 
     // Disallow copying and assigning.
     Utf8Value(const Utf8Value&) = delete;
@@ -588,7 +588,7 @@ class V8_EXPORT String : public Name {
     ~Value();
     uint16_t* operator*() { return str_; }
     const uint16_t* operator*() const { return str_; }
-    int length() const { return length_; }
+    uint32_t length() const { return length_; }
 
     // Disallow copying and assigning.
     Value(const Value&) = delete;
@@ -625,7 +625,7 @@ class V8_EXPORT String : public Name {
 #endif
       return data16_;
     }
-    int length() const { return length_; }
+    uint32_t length() const { return length_; }
     bool is_one_byte() const { return is_one_byte_; }
 
     // Disallow copying and assigning.
@@ -640,7 +640,7 @@ class V8_EXPORT String : public Name {
       const uint8_t* data8_;
       const uint16_t* data16_;
     };
-    int length_;
+    uint32_t length_;
     bool is_one_byte_;
     // Avoid exposing the internal DisallowGarbageCollection scope.
     alignas(internal::Internals::
